Fixing dictionary year parsing
Existing code was only returning years 1972 to 2015, which resulted
in earlier data being entirely excluded. I think this has something
to do with the way that I was looping through the json nodes, but
in any case we can rely upon the years defined in the record check
property node as our master list of years. Eventually this should
be fixed by someone smarter than me.