Data raw process fix (#17)
* Fixing dictionary year parsing

Existing code was only returning years 1972 to 2015, which resulted
in earlier data being entirely excluded. I think this has something
to do with the way that I was looping through the json nodes, but
in any case we can rely upon the years defined in the record check
property node as our master list of years. Eventually this should
be fixed by someone smarter than me.

* I'm a moron

Removing the evidence of that.

* Adding corrected births data

This reincorporates the 1968-1971 data that was lost due to dictionary parsing errors.