Doc corrections (#19)
* Correcting main README references

* Updating unzip tools reference

Since we changed from linux unzip to 7zip

* Updated column definition names

Matching the column name exactly allows the user to search for the column name as-is, which also allows hotkeys to be used on column references. This makes the documentation of each column much easier to access.

* Updating version an suggestions

* Correcting dependency

There are still dplyr based functions in the package. These will eventually be removed, or replaced by dtplyr based functions. For now, correcting the reference so that the build can function.