Birth in hospital docs (#21)
* Adding column type tests

* Adding attendant at birth columns

We have no data on whether births occurred in hospital prior to
1975 without adding this field. Since we're performing a simple
logical of hospital True/False, this field should suffice to give
us data for all years on births in or out of hospital

* Trying to add 68-78 birth in hospital; debug

* Initial take on definition and tests

These still aren't final because the R reduction processing script is still
broken. That will need to be fixed before these other components are
finalized.

* Converting from 4 spaces to 2

Need more room for code

* Adding handling of birth_in_hospital sources

* Re-enabling 2015 as default end year

This config file needs to change eventually, but for now I'm just
re-establishing it for the complete range of years.

* Adding births data with corrections

The corrections that allow us to identify whether a birth occured
in the hospital were processed successfully, and the data set is
passing the new tests that were written. Documentation is still
par-baked though.

* Updated description