Updating dictionary value
One of the mapped values included a single quote ('), which when
used as a list value mapping in R would cause SublimeText to parse
the rest of the script incorrectly. This didn't cause direct issues
in the processing, but it made the script annoying to work with
after that. Fixing this.