Adding age recode for 2003
An age recode field was used for a single year. It appears that the
dictionary had a bit of a hiccup in 03, as the DMAGE field ended in 2003,
but the succeeding MAGER field didn't start until 2004. This field was
apparently the stop-gap. Mapping it in, but 2003 will have a little less
completeness for age than the other years.