Disabling sampling in conofig by default
I'm planning to eventually have the server do all the processing all the time
so I don't need to clunky config file, which really adds a lot more problems
than it fixes. Evenually it will be phased out entirely, but for now I'm
setting the confuration so that I don't have to change it on the server
with every push.