Needed to explicitly cast factor correction as int
The formula that adds to the MAGER41 factor was causing an error, because
R interpreted the value as a double instead of an int. This caused R to
throw an error in the cases where MAGER41 existed, but that was not seen
because try-catch stuff was happening and preventing it. Overall, this raw
processing script is still a big mess.