Adding further scaffolding for documentation
Mother Age Labeled has been completely replaced by the mother age int column,
which meant the removal of some references to the old field. Additionally, the
births data set has been moved into it's out file for documentation, references
to the columns are all provided as links to the main page, and the formatting
of the entire births documentation is generally improving.

A stub file for testing mother age has been added, but currently will fail any
testing (on purpose).