Mother age docs (#23)
* Renaming tests for consistency with definitions

* Adding definitions for imputed age

* Adding mapping to reduction script

* Initial draft of test

* Fixing recode field references

Screwed this up in a different branch, but didn't notice because
I selectively rebuilt the year files.

* Updating dictionary value

One of the mapped values included a single quote ('), which when
used as a list value mapping in R would cause SublimeText to parse
the rest of the script incorrectly. This didn't cause direct issues
in the processing, but it made the script annoying to work with
after that. Fixing this.

* Playing with maternal age remapping

Still getting mostly missing data from 92-02. Trying to fix that.

* Discovered that 1993 was missing from age mapping

* Adding age recode for 2003

An age recode field was used for a single year. It appears that the
dictionary had a bit of a hiccup in 03, as the DMAGE field ended in 2003,
but the succeeding MAGER field didn't start until 2004. This field was
apparently the stop-gap. Mapping it in, but 2003 will have a little less
completeness for age than the other years.

* Adding processing rules for MAGER41

This one-time refactoring of age for 2003 has a unusual schema
compared to all other years. Adding a rule for handling this data
in the hopes that this will help with the missing data in 03.

* Correcting MAGER41 formula

* Documenting mother_age_int and removing labeled

Getting rid fo the labeled maternal age. Although we'll lose some
granularity for births at the extremes of the ranges (less than
13 or 15, greater than 50), this is a fairly minor loss of data
and it will greatly simplify the dataset as a whole.

* Renaming mother_age_int to mother_age

Taking the place of the vacated labeled age field.

* Disabling sampling in conofig by default

I'm planning to eventually have the server do all the processing all the time
so I don't need to clunky config file, which really adds a lot more problems
than it fixes. Evenually it will be phased out entirely, but for now I'm
setting the confuration so that I don't have to change it on the server
with every push.

* Needed to explicitly cast factor correction as int

The formula that adds to the MAGER41 factor was causing an error, because
R interpreted the value as a double instead of an int. This caused R to
throw an error in the cases where MAGER41 existed, but that was not seen
because try-catch stuff was happening and preventing it. Overall, this raw
processing script is still a big mess.

* Adding further scaffolding for documentation

Mother Age Labeled has been completely replaced by the mother age int column,
which meant the removal of some references to the old field. Additionally, the
births data set has been moved into it's out file for documentation, references
to the columns are all provided as links to the main page, and the formatting
of the entire births documentation is generally improving.

A stub file for testing mother age has been added, but currently will fail any
testing (on purpose).

* Adding tests and data set